fix issue: 【BUG】Traffic tracking seems to have stopped working. #115

+++ app/src/main/java/com/android/xrayfa/TrafficDetector.kt
@@ -1,6 +1,5 @@
 package com.android.xrayfa
 
-import kotlinx.coroutines.CoroutineScope
 
 /**
  * Traffic detector, used to calculate upload and download speeds for front-end display
@@ -12,6 +11,6 @@ interface TrafficDetector {
     fun stopTrafficDetection()
 
 
-    suspend fun consumeTraffic(onConsume: suspend (Pair<Long, Long>) -> Unit)
+    suspend fun consumeTraffic(onConsume: suspend (Pair<Double, Double>) -> Unit)
 
 }
